Extract FilePane into file_manager.pane
Move the FilePane class (~2500 lines) from file_manager_window.py into
its own file_manager/pane.py submodule.  Update __init__.py to re-export
FilePane and its icon-size constants, and add the missing FilePane import
in file_manager_window.py.
