+++ sshpilot/file_manager/__init__.py
@@ -1,18 +1,26 @@
 """File-manager subpackage.
 
 Step 4 of the refactor plan extracts the monolithic ``file_manager_window``
-module into focused submodules. Phase 4a moved the low-coupling helpers
+module into focused submodules.  Phase 4a moved the low-coupling helpers
 (portal/docs handling, format helpers, paramiko walk helpers, the
-cancellation exception). Phase 4b (this commit) moves the standalone
-dialogs and pane-level UI controls (``SFTPProgressDialog``, the
-``PathEntry``/``PaneControls``/``PaneToolbar`` pane chrome, and
-``PropertiesDialog``). The remaining heavy widget classes
-(``AsyncSFTPManager``, ``FilePane``, ``FileManagerWindow``) still live
-in ``sshpilot.file_manager_window`` and will move in 4c.
+cancellation exception).  Phase 4b moved the standalone dialogs and
+pane-level UI controls (``SFTPProgressDialog``, ``PathEntry``,
+``PaneControls``, ``PaneToolbar``, ``PropertiesDialog``).  Phase 4c
+extracted the heavy widget classes: ``AsyncSFTPManager`` (4c-i) and
+``FilePane`` (4c-ii).  ``FileManagerWindow`` remains in
+``sshpilot.file_manager_window`` and will move in 4c-iii.
 """
 
 from .exceptions import TransferCancelledException
 from .format_utils import _human_size, _human_time, _mode_to_octal, _mode_to_str
+from .pane import (
+    FilePane,
+    _DEFAULT_ICON_LEVEL,
+    _GRID_ICON_SIZES,
+    _LIST_ICON_SIZES,
+    _MAX_ICON_LEVEL,
+    _MIN_ICON_LEVEL,
+)
 from .pane_controls import PaneControls, PaneToolbar, PathEntry
 from .sftp_manager import AsyncSFTPManager, FileEntry, _MainThreadDispatcher
 from .portal_docs import (
